BlogPageListItem componente
Summary:

Test Plan:
