+++ packages/client/components/blog/BlogPageContent.tsx
@@ -2,9 +2,9 @@ import { useFragment, useLazyLoadQuery } from "react-relay";
 import { getLogger } from "deps.ts";
 import { graphql } from "packages/client/deps.ts";
 import type { BlogPostContentFragment$key } from "packages/__generated__/BlogPostContentFragment.graphql.ts";
-import { BfSymbol } from "packages/bfDs/static/BfSymbol.tsx";
 import { BfDsButton } from "packages/bfDs/BfDsButton.tsx";
 import { BlogPageHero } from "packages/client/components/blog/BlogPageHero.tsx";
+import { BlogPageListItem } from "packages/client/components/blog/BlogPageListItem.tsx";
 
 const logger = getLogger(import.meta);
 const fragment = await graphql`
@@ -66,52 +66,7 @@ export function BlogPageContent({ blogRef }: Props) {
     }
   });
 
-  return blogPostList;
-}
-
-function BlogPageListItem({ postRef, additionalClassName }: PostProps) {
-  const post = useFragment(fragment, postRef);
-  const { navigate } = useRouter();
-  let coverImage;
-  if (post.coverUrl) {
-    coverImage = (
-      <img
-        className="blog-list-item-image"
-        src={post.coverUrl}
-        alt="Cover Image"
-      />
-    );
-  } else {
-    coverImage = (
-      <div className="blog-list-item-image-placeholder blog-list-item-image-placeholder-fourthary">
-        <div style={{ width: "50%" }}>
-          <BfSymbol color="white" />
-        </div>
-      </div>
-    );
-  }
-  return (
-    <div
-      key={post.id}
-      className={`blog-page-list-item ${additionalClassName || ""}`}
-      onClick={() => navigate(`/blog/${post.slug}`)}
-    >
-      <div className="blog-page-item-info">
-        <div className="blog-list-item-title">{post.title}</div>
-        <div className="blog_author">
-          <div
-            className="blog_author_avatar"
-            style={{ backgroundImage: `url(${post?.author?.avatarUrl})` }}
-          />
-          <div>{post?.author?.name}</div>
-        </div>
-        {/* <div>{post?.author?.email}</div> */}
-        <div className="blog_meta">{post?.date}</div>
-        <div className="blog-list-item-summary">{post.summary}</div>
-      </div>
-      {coverImage}
-    </div>
-  );
+  return <>{blogPostList}</>;
 }
 
 function BlogPageCta() {
